Rename kai-functions -> labs-functions
