+++ .nxignore
@@ -4,7 +4,7 @@
 packages/experimental/kai-types/**
 packages/experimental/kai/**
 packages/experimental/kai-frames/**
-packages/experimental/kai-functions/**
+packages/experimental/labs-functions/**
 packages/experimental/kai-bots/**
 packages/experimental/kai-framework/**
 packages/experimental/kube-console/**
