+++ packages/experimental/labs-functions/functions/chess/package.json
