+++ packages/experimental/labs-functions/functions/chess/tsconfig.json
