+++ packages/experimental/labs-functions/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dxos/kai-functions",
+  "name": "@dxos/labs-functions",
   "version": "0.3.0",
   "private": true,
   "description": "Experimental functions.",
