+++ packages/experimental/labs-functions/scripts/clean.sh
