+++ packages/experimental/labs-functions/scripts/deploy.sh
