+++ packages/experimental/labs-functions/scripts/env.sh
