+++ packages/experimental/labs-functions/src/functions/chatgpt/chat-model.ts
