+++ packages/experimental/labs-functions/src/functions/chatgpt/handler.ts
