+++ packages/experimental/labs-functions/src/functions/chatgpt/index.ts
