+++ packages/experimental/labs-functions/src/functions/chatgpt/parser.ts
