+++ packages/experimental/labs-functions/src/util.ts
