+++ pnpm-workspace.yaml
@@ -2,11 +2,11 @@ packages:
   - docs
   - packages/**/*
   - tools/**/*
-  - '!packages/experimental/kai-functions/functions/**'
+  - '!packages/experimental/labs-functions/functions/**'
   - '!packages/experimental/kai-types'
   - '!packages/experimental/kai'
   - '!packages/experimental/kai-frames'
-  # - '!packages/experimental/kai-functions'
+  # - '!packages/experimental/labs-functions'
   - '!packages/experimental/kai-bots'
   - '!packages/experimental/kai-framework'
   - '!packages/experimental/kube-console'
