+++ release-please-config.json
@@ -499,17 +499,17 @@
         },
         {
           "type": "json",
-          "path": "packages/experimental/kai-functions/package.json",
-          "jsonpath": "$.version"
-        },
-        {
-          "type": "json",
           "path": "packages/experimental/kai-sandbox/package.json",
           "jsonpath": "$.version"
         },
         {
           "type": "json",
           "path": "packages/experimental/kodama/package.json",
+          "jsonpath": "$.version"
+        },
+        {
+          "type": "json",
+          "path": "packages/experimental/labs-functions/package.json",
           "jsonpath": "$.version"
         },
         {
